// src/lib/integrations/google/models/google-drive-folder.ts
/**
 * ============================================
 * CLARA OS
 * Google Drive Integration
 * --------------------------------------------
 * File : google-drive-folder.ts
 * ============================================
 */

export interface GoogleDriveFolder {

  id: string;

  name: string;

  parentId?: string;

}